+++ Security/db_dsmrapi.sql
@@ -9,3 +9,5 @@ GRANT EXECUTE ON [dbo].[DsmrApi_SelectUserById] TO [db_dsmrapi]
 GO
 GRANT EXECUTE ON [dbo].[DsmrApi_SelectDevicesForUser] TO [db_dsmrapi]
 GO
+GRANT EXECUTE ON [dbo].[DsmrApi_Login] TO [db_dsmrapi]
+GO