+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 18-04-2021
+### Added
+- Procedure to allow users to login
+- Device onboarding tokens
+
+### Updated
+- DSMR API database role
+- Devices table
+
 ## [1.2.0] - 18-04-2021
 ### Added
 - Onboarding token to the user table